Refactoring part of code
